fix: cast share card bottom coordinate to double
Agent-Logs-Url: https://github.com/ivansanare93/Bolometro/sessions/2d33b937-e132-4df8-99c4-5868defad379

+++ lib/utils/compartir_sesion.dart
@@ -583,7 +583,7 @@ Future<Uint8List> _renderShareCard(_ShareCardContent content) async {
 
   top += 36;
   final detailPanelTop = top - 10;
-  final detailPanelBottom = content.notes == null ? height - 190 : height - 250;
+  final detailPanelBottom = (content.notes == null ? height - 190 : height - 250).toDouble();
   final detailPanel = RRect.fromRectAndRadius(
     Rect.fromLTRB(
       horizontalPadding - 6,
